Google scholars/requirements.txt
alabaster==0.7.12
arrow==1.2.2
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
Automat==20.2.0
Babel==2.9.1
beautifulsoup4==4.10.0
bibtexparser==1.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
Deprecated==1.2.13
django-cors-headers==3.11.0
django-truncate==0.1
djangorestframework==3.13.1
docutils==0.17.1
fake-useragent==0.1.11
free-proxy==1.0.6
future==0.18.2
h11==0.13.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
Jinja2==3.0.3
jmespath==1.0.0
json5==0.9.6
lxml==4.8.0
MarkupSafe==2.1.1
outcome==1.1.0
packaging==21.3
parsel==1.6.0
priority==1.3.0
Protego==0.2.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
queuelib==1.6.2
requests==2.27.1
scholarly==1.6.0
scrapeops-scrapy==0.4.6
Scrapy==2.5.1
selenium==4.1.3
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
Sphinx==4.4.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
tld==0.12.6
trio==0.20.0
trio-websocket==0.9.2
Twisted==22.2.0
typing_extensions==4.1.1
URLEncoder==1.1
urllib3==1.26.8
w3lib==1.22.0
wrapt==1.14.0
wsproto==1.1.0
zipp==3.7.0
zope.interface==5.4.0